cell-instance-segmentation/requirements_cluster.txt
absl-py==0.9.0
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
cloudpickle==1.4.1
cycler==0.10.0
Cython==0.29.19
detectron2==0.1.3+cpu
future==0.18.2
fvcore==0.1.1.post20200608
google-auth==1.16.1
google-auth-oauthlib==0.4.1
googledrivedownloader==0.4
grpcio==1.29.0
idna==2.9
importlib-metadata==1.6.1
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
mock==4.0.2
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.2.0.34
Pillow==7.1.2
pkg-resources==0.0.0
portalocker==1.7.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.0
pydot==1.4.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
tabulate==0.8.7
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
termcolor==1.1.0
torch==1.5.0
torchvision==0.6.0
tqdm==4.46.1
urllib3==1.25.9
Werkzeug==1.0.1
yacs==0.1.7
zipp==3.1.0
